+++ Changelog.md
@@ -1,5 +1,10 @@
 # Factureaza PHP SDK Changelog
 
+## Unreleased
+##### 2022-XX-YY
+
+- Added retrieve invoice PDF support
+
 ## 1.2.0
 ##### 2022-10-11